backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
pdfplumber==0.10.3
python-docx==0.8.11
reportlab==4.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
